PR-2356 Working version of PointCloud
Strip out the worst of the PointMatching-specific functionality, and
start refactoring into a generic point cloud data structure.  Move to
the convention that point attribute columns are Vector{T} rather than
general arrays with the last dimension indexing the points.  In line
with this, start using FixedSizeArrays.Vec{3} for position and normal
information.